Reach the rainbow functionality
Completed win if you reach the rainbow function, advance-stage function, implemented super-powers with collision detection.
Had to change the world struct to include the stage, and that meant updating all the pad-handler options. Added a let* to many of the functions to increase readability.
Reorganized some of the code:put structs and functions together, images together, etc,
Added comments.